requirements.txt
Flask==3.0.0
pymongo==4.6.1
dnspython==2.4.2
Werkzeug==3.0.1
python-dotenv==1.0.1
langchain-groq==0.0.6
langchain==0.1.0
certifi==2024.2.2
